Move doc-comment utilities from shared/ to transforms/doc-comment/
Moved doc-comment-description.ts and doc-comment-normalization.ts from
src/plugin/src/shared/ to src/plugin/src/transforms/doc-comment/ as
description-utils.ts and normalization-utils.ts respectively.

Rationale:
- The shared/ directory contained only doc-comment specific utilities
- These utilities are exclusively used by doc-comment transform and printer code
- Moving them to transforms/doc-comment/ co-locates related functionality
- Maintains architectural boundaries: all doc-comment code now lives together
- Removes the unnecessary shared/ directory which had no other purpose

Updated imports in:
- src/plugin/src/transforms/doc-comment/doc-comment-normalization.ts
- src/plugin/src/printer/doc-comment/function-docs.ts
- src/plugin/src/transforms/doc-comment/index.ts (barrel exports)

Build and tests verified to pass.
