--- src/plugin/src/shared/index.ts
@@ -1,2 +1,0 @@
-export * as DocCommentDescription from "./doc-comment-description.js";
-export * as DocCommentNormalization from "./doc-comment-normalization.js";
