+++ src/plugin/src/transforms/doc-comment/description-utils.ts
