+++ src/plugin/src/transforms/doc-comment/index.ts
@@ -6,3 +6,5 @@ export {
     getDeprecatedDocCommentFunctionSet
 } from "./doc-comment-metadata.js";
 export { docCommentNormalizationTransform } from "./doc-comment-normalization.js";
+export * as DescriptionUtils from "./description-utils.js";
+export * as NormalizationUtils from "./normalization-utils.js";
