+++ src/plugin/src/transforms/doc-comment/normalization-utils.ts
